Clear the Search query when ChordSelector clicked.
If there's some text in the search query, we want to clear it when the
chord selector selects a chord. Otherwise we may have mismatch values in
the chord selector and in the search query string.

I'm using a weird boolean toggle in the App state. I don't know of a
better way, but there must be something better. Basically I want to send
a message from one inner component to another inner component. Perhaps I
need to a common value that is used by ChordSelector's Message, and
Search's Input.