Store lookahead chord so we can revert to original.
When a user types a search query, we want to change the active fretboard
to display the first result of the query. BUT when the user clears the
query, we want to revert to whatever chord was *last* manually-selected
by the user.

To do this, we create a new state attribute in App, called
lookaheadChord. This chord, if not empty, elides the current chord. But
when a user selects a chord, we erase the lookahead chord.